Move donate lektor packages to be local
Adding the packages the new donate stuff is using to the repository so
that they do not need to be downloaded to run lektor. I had to make some
modifications to the setup.py files for these modules in order to get
them to work. Otherwise I was getting errors about invalid syntax in the
requires.txt files and also one of the modules was requiring lektor
which was pulling in a bunch of stuff that's not needed because we are
running inside lektor. I put all these changes in patch files in a new
patches subdirectory.