fontawesome cleanup
We really only use a small portion of the FontAwesome package, so we
should only keep the parts that we need. This removes over 5000 files
that are copied every time we build the sites locally and in CI so this
should help performance in both build and deployment processes.

This keeps around a handful of files in png/* that are referred to by
bootstrap.css as well and the CSS code and the webfont files.