Add lektor-scss plugin
Closes: #27

I've audited the plugin, and it's safe. it's a straight lektor plugin wrapper
around libsass, with most of the logic handling dev server reloads.

Each site individually will have to switch to the plugin to compile its
SCSS. This will require a lego update, a new config file, and a change
to the CI script.