persist yec banner close during navigation
While the checkbox persistence mechanism is clever for not requiring
javascript, it only really works for reloads of the *same* page/url. The
checkbox only remembers its state on a given page, not all pages of the
same site. For example, even if the visitor closes the banner on the front
page, it will be shown when navigating to /about, and so forth.

This leverages the browser's sessionStorage to remember when the banner
is closed, and persists for all pages of the same website until the browser
session or tab is closed, which should be much more convenient for casual
visitors.